requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
django-rest-framework>=0.1.0
djangorestframework>=3.12.4
django-cors-headers>=3.8.0
django-filter